+++ app/docs/button.docs.tsx
@@ -1,20 +1,28 @@
 import { Button } from "theme-ui";
 import { markdown, ReactSpecimen } from "catalog";
+import { Icon } from "../icons";
+import SvgIcChevronRight from "../icons/components/IcChevronRight";
+import SvgIcChevronLeft from "../icons/components/IcChevronLeft";
 
 export default () => markdown`
 > Buttons are used to trigger an event after a user interaction.
 
-There are four basic styles that are styles defined in \`rebass\`'s \`variants\`:
+There are four basic styles that are styles defined in \`theme-ui\`'s \`variants\`:
 
 - \`primary\`
+- \`primary-small\`
 - \`secondary\`
 - \`success\`
-- \`outline\`
+- \`inline\`
+- \`inline-bold\`
 
 
   ${(
     <ReactSpecimen span={2}>
-      <Button variant="primary">Primary button</Button>
+      <Button variant="primary">
+        <Icon name="linkExternal" />
+        <span>Primary button</span>
+      </Button>
     </ReactSpecimen>
   )}
 
@@ -36,11 +44,37 @@ There are four basic styles that are styles defined in \`rebass\`'s \`variants\`
       <Button variant="outline">Outline button</Button>
     </ReactSpecimen>
   )}
+  ${(
+    <ReactSpecimen span={2}>
+      <Button variant="inline">
+        <SvgIcChevronLeft />
+        <span>Inline button</span>
+      </Button>
+    </ReactSpecimen>
+  )}
+  ${(
+    <ReactSpecimen span={2}>
+      <Button variant="inline-bold">
+        <SvgIcChevronLeft />
+        Bold inline button
+      </Button>
+    </ReactSpecimen>
+  )}
+  ${(
+    <ReactSpecimen span={2}>
+      <Button variant="primary-small">
+        <span>Publish this dataset</span>
+        <SvgIcChevronRight />
+      </Button>
+    </ReactSpecimen>
+  )}
 
   ## How to use
 
 ~~~
 import { Button } from "theme-ui"
+import SvgIcChevronRight from "../icons/components/IcChevronRight";
+import SvgIcChevronLeft from "../icons/components/IcChevronLeft";
 
 <Button variant="primary">
   Primary button